pyo3-build-config: new crate to re-use build.rs across crates
