+++ Cargo.toml
@@ -11,7 +11,6 @@ documentation = "https://docs.rs/crate/pyo3/"
 categories = ["api-bindings", "development-tools::ffi"]
 license = "Apache-2.0"
 exclude = ["/.gitignore", ".cargo/config", "/codecov.yml", "/Makefile", "/pyproject.toml", "/tox.ini"]
-build = "build.rs"
 edition = "2018"
 
 [dependencies]
@@ -40,6 +39,9 @@ proptest = { version = "0.10.1", default-features = false, features = ["std"] }
 pyo3 = { path = ".", default-features = false, features = ["macros", "auto-initialize"] }
 serde_json = "1.0.61"
 
+[build-dependencies]
+pyo3-build-config = { path = "pyo3-build-config", version = "=0.14.0-alpha.0" }
+
 [features]
 default = ["macros"]
 
@@ -55,13 +57,13 @@ multiple-pymethods = ["inventory"]
 extension-module = []
 
 # Use the Python limited API. See https://www.python.org/dev/peps/pep-0384/ for more.
-abi3 = []
+abi3 = ["pyo3-build-config/abi3"]
 
 # With abi3, we can manually set the minimum Python version.
-abi3-py36 = ["abi3-py37"]
-abi3-py37 = ["abi3-py38"]
-abi3-py38 = ["abi3-py39"]
-abi3-py39 = ["abi3"]
+abi3-py36 = ["abi3-py37", "pyo3-build-config/abi3-py36"]
+abi3-py37 = ["abi3-py38", "pyo3-build-config/abi3-py37"]
+abi3-py38 = ["abi3-py39", "pyo3-build-config/abi3-py38"]
+abi3-py39 = ["abi3", "pyo3-build-config/abi3-py39"]
 
 # Changes `Python::with_gil` and `Python::acquire_gil` to automatically initialize the
 # Python interpreter if needed.
