+++ examples/pyo3-pytests/Cargo.toml
@@ -6,10 +6,10 @@ description = "Python-based tests for PyO3"
 edition = "2018"
 
 [dependencies]
+pyo3 = { path = "../../", features = ["extension-module"] }
 
-[dependencies.pyo3]
-path = "../../"
-features = ["extension-module"]
+[build-dependencies]
+pyo3-build-config = { path = "../../pyo3-build-config" }
 
 [lib]
 name = "pyo3_pytests"
